move coders into db package
