+++ apps/api/package.json
@@ -74,7 +74,6 @@
   "dependencies": {
     "@privy-io/server-auth": "^1.31.0",
     "@envio-dev/hypersync-client": "^0.6.5",
-    "@recallnet/coders": "workspace:*",
     "@recallnet/db-schema": "workspace:*",
     "@recallnet/staking-contracts": "workspace:*",
     "@sentry/node": "^10.5.0",
