--- packages/coders/eslint.config.js
@@ -1,4 +1,0 @@
-import { config } from "@recallnet/eslint-config/base";
-
-/** @type {import("eslint").Linter.Config} */
-export default config;
