--- packages/coders/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@recallnet/coders",
-  "version": "0.0.0",
-  "license": "MIT AND Apache-2.0",
-  "private": true,
-  "description": "Recall custom encoders and decoders.",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/recallnet/js-recall.git",
-    "directory": "packages/coders"
-  },
-  "homepage": "https://github.com/recallnet/js-recall#readme",
-  "type": "module",
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      },
-      "require": {
-        "types": "./dist/index.d.cts",
-        "default": "./dist/index.cjs"
-      }
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "dev": "tsc --watch",
-    "build": "tsup",
-    "clean": "rm -rf .turbo node_modules dist",
-    "lint": "eslint . --max-warnings 0",
-    "docs:check": "typedoc --emit none",
-    "docs:build": "typedoc",
-    "docs:watch": "typedoc --watch",
-    "docs:serve": "typedoc --watch --serve"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "@recallnet/eslint-config": "workspace:*",
-    "@recallnet/typescript-config": "workspace:*",
-    "@types/node": "^24.3.0",
-    "tsup": "^8.3.6",
-    "typedoc": "^0.28.1",
-    "typedoc-plugin-coverage": "^3.4.1",
-    "typedoc-plugin-markdown": "^4.6.0",
-    "typescript": "^5.7.3"
-  }
-}
