--- packages/coders/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "@recallnet/typescript-config/base.json",
-  "compilerOptions": {
-    "outDir": "dist"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules"]
-}
