--- packages/coders/tsup.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig({
-  entry: ["src/index.ts"],
-  format: ["cjs", "esm"],
-  dts: true,
-  clean: true,
-  outExtension: ({ format }) => ({
-    js: format === "esm" ? ".js" : ".cjs",
-  }),
-});
